% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ten_to_mean.R
\name{add_ten_to_mean}
\alias{add_ten_to_mean}
\title{Add Ten to Mean}
\usage{
add_ten_to_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value that is the mean of \code{x} plus 10.
}
\description{
This function calculates the mean of a numeric vector and adds 10 to it.
}
\examples{
add_ten_to_mean(c(1, 2, 3, 4, 5))
}
