% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\docType{data}
\name{cell_colors}
\alias{cell_colors}
\title{colors from cell}
\format{
An object of class \code{character} of length 64.
}
\usage{
cell_colors
}
\description{
colors from cell
}
\author{
Kai Guo
}
\keyword{datasets}
