% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{col2hex}
\usage{
col2hex(col = "grey80")
}
\arguments{
\item{col}{character, color name}
}
\value{
character, hex color
}
\description{
col2hex
}
\examples{
col2hex("grey80")
}
