% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{colors_group}
\alias{colors_group}
\title{270 ncl colors}
\format{
An object of class \code{list} of length 34.
}
\usage{
colors_group
}
\description{
-\code{rcolors}: list of 270 colors
\itemize{
\item \code{colors_group}: list of grouped colors
}
}
\examples{
str(rcolors[1:6])
str(colors_group[1:6])
}
\references{
https://www.ncl.ucar.edu/Document/Graphics/color_table_gallery.shtml
}
\keyword{internal}
