% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\docType{data}
\name{distcolor}
\alias{distcolor}
\title{distinguish colors for making figures}
\format{
An object of class \code{character} of length 61.
}
\usage{
distcolor
}
\description{
distinguish colors for making figures
}
\author{
Kai Guo
}
\keyword{datasets}
