% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{gcolor}
\alias{gcolor}
\title{generate random colors without grey o not}
\usage{
gcolor(ncolor, grey = FALSE, alpha = 1)
}
\arguments{
\item{grey}{include the grey color or not (default: FALSE)}

\item{alpha}{Transparent level
author Kai Guo}
}
\description{
generate random colors without grey o not
}
