% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_colors_by_category}
\alias{get_colors_by_category}
\title{获取指定类别的颜色}
\usage{
get_colors_by_category(category, n = NULL)
}
\arguments{
\item{category}{颜色类别，例如 "Red", "Green", "Blue", "Neutral" 等}

\item{n}{所需颜色数量}
}
\value{
返回颜色代码的向量
}
\description{
获取指定类别的颜色
}
