% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{hex2HSL}
\alias{hex2HSL}
\title{将十六进制颜色代码转换为 HSL 值}
\usage{
hex2HSL(hex)
}
\arguments{
\item{hex}{字符向量，包含十六进制颜色代码（可带或不带 "#" 前缀）}
}
\value{
返回一个数据框，包含 \code{H}, \code{S}, \code{L} 三列，分别为色相、饱和度和亮度
}
\description{
该函数将十六进制颜色代码转换为 HSL（色相、饱和度、亮度）颜色空间的值。
}
\examples{
hex_colors <- c("FF0000", "#00FF00", "0000FF")
hsl_values <- hex2HSL(hex_colors)
print(hsl_values)
}
