% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcolors-package.R
\docType{package}
\name{rcolors-package}
\alias{rcolors}
\alias{rcolors-package}
\title{rcolors: NCL color tables in R language}
\description{
NCL color tables in R language.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
