% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cols.R
\name{show_cols}
\alias{show_cols}
\alias{show_cols_brewer}
\alias{show_col}
\title{Show list of colors}
\usage{
show_cols(colors_list, margin = 8, fontsize = 1, family = NULL)

show_cols_brewer(colors_list)

show_col(colors, labels = TRUE, borders = NULL, cex_label = 1, nrow = NULL)
}
\arguments{
\item{colors_list}{A list objects filled with colors_list or a color vector}

\item{margin}{margin in the left}

\item{family, fontsize}{font family and size of color names in the left}

\item{colors}{a character vector of colors}

\item{labels}{boolean, whether to show the hexadecimal representation of the colours in each tile}

\item{borders}{colour of the borders of the tiles}

\item{cex_label}{size of printed labels, works the same as cex parameter of \code{\link[=plot]{plot()}}}

\item{nrow}{integer, number of rows}
}
\description{
Show list of colors

\itemize{
\item \code{show_cols}: show list of colors
\item \code{show_cols_brewer}: show list of colors in RColorBrewer style
}
}
\examples{
print(names(colors_group))
show_cols(colors_group$rainbow, margin = 14)
}
