% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{Transparent}
\alias{Transparent}
\title{Transparent color}
\usage{
Transparent(colors, alpha = 0.8)
}
\arguments{
\item{colors}{a vector of colors: "red","blue"}

\item{alpha}{Transparent level}
}
\description{
Transparent color
}
\examples{
{
Transparent(2, 4)
Transparent("red", "blue")
Transparent(rgb(1,0,0), rgb(0,0,1))
}
}
\author{
Kai Guo
}
